% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ALL}
\alias{ALL}
\title{ALL data from the ALL package}
\format{ExpressionSet with 12625 features, 128 samples}
\source{
The \code{ALL} R package
}
\usage{
ALL
}
\description{
This object is a copy of \code{\link[ALL]{ALL}}
}
\keyword{datasets}
