% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{biogrid2011}
\alias{biogrid2011}
\title{BioGRID PPI network until 2011}
\format{An \code{\link[igraph]{igraph}} object with entrez ids as vertex names and
confidence as edge weights}
\source{
Database from \url{https://thebiogrid.org/} and
edge weights from [Cao et al, 2014]
  \url{http://dx.doi.org/10.1093/bioinformatics/btu263}
}
\usage{
biogrid2011
}
\description{
A protein-protein interaction network containing 11,394
nodes and 67,573 edges. The publications to derive the
interactions to build the nework
are up to the year 2011, not included. The network is weighted according to
the confidence on the interactions [Cao et al, 2014]
  \url{http://dx.doi.org/10.1093/bioinformatics/btu263}, mainly computed from the
number of publications supporting it and their throughput, being low
throughput regarded as more trustable.
}
\keyword{datasets}
