% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataLym}
\alias{dataLym}
\title{dataLym gene expression p-values from the DLBCL package}
\format{A data.frame with 3583 rows and 8 columns, with genes
and p-values for differential expression.}
\source{
The \code{DLBCL} R package
}
\usage{
dataLym
}
\description{
This object is a copy of \code{\link[DLBCL]{dataLym}}
}
\keyword{datasets}
