% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{interactome}
\alias{interactome}
\title{HPRD network from the DLBCL package}
\format{A graphNEL object with 9386 nodes and 36504 edges, representing a
human interactome from HPRD.}
\source{
The \code{DLBCL} R package
}
\usage{
interactome
}
\description{
This object is a copy of \code{\link[DLBCL]{interactome}}
}
\keyword{datasets}
