% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kegg2011}
\alias{kegg2011}
\title{KEGG pathways from 2011}
\format{A list with 123 KEGG pathways}
\source{
The outdated bioconductor repository KEGG.db
\url{http://www.bioconductor.org/packages/release/data/annotation/html/KEGG.db.html}
}
\usage{
kegg2011
}
\description{
123 KEGG pathways from 14th March, 2011. These pathways
contain entrez ids and are guaranteed to have new genes
in \link[retroData]{kegg2017} and to perfectly
map to the \link[retroData]{biogrid2011}
network.
}
\keyword{datasets}
