% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kegg2017}
\alias{kegg2017}
\title{KEGG pathways from 2017}
\format{A list with 123 KEGG pathways}
\source{
The KEGG rest API \url{rest.kegg.jp}
}
\usage{
kegg2017
}
\description{
123 KEGG pathways from 17th January, 2017. These pathways
contain entrez ids and are guaranteed to have new genes
compared to \link[retroData]{kegg2011} and to
perfectly map to the \link[retroData]{biogrid2011}
network.
}
\keyword{datasets}
