% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kegg2018}
\alias{kegg2018}
\title{KEGG pathways from 2018}
\format{A list with 123 KEGG pathways}
\source{
The KEGG rest API \url{rest.kegg.jp}
}
\usage{
kegg2018
}
\description{
139 KEGG pathways from 18th August 2018. These pathways
contain entrez ids and are guaranteed to have new genes
compared to \link[retroData]{kegg2011} and to
perfectly map to the \link[retroData]{biogrid2011}
network.
}
\keyword{datasets}
