% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{keggdlbcl}
\alias{keggdlbcl}
\title{KEGG pathways for the DLBCL/HPRD simulated gene expression}
\format{A table with 28077 rows, 2 columns, with 320 KEGG pathways}
\source{
The KEGG rest API \url{rest.kegg.jp}
}
\usage{
keggdlbcl
}
\description{
320 KEGG pathways from 2017, Release 83.0+/09-09, Sep 17.
}
\keyword{datasets}
