% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yeast}
\alias{yeast}
\title{Yeast network from the igraphdata package}
\format{An igraph object with 2617 nodes and 11855 edges, representing a
yeast protein interactions, von Mering et al.}
\source{
The \code{igraphdata} R package
}
\usage{
yeast
}
\description{
This object is a copy of \code{\link[igraphdata]{yeast}}
}
\keyword{datasets}
